fix all patters for my 300 leds strip

+++ GradientPalettes.h
@@ -513,6 +513,6 @@ const TProgmemRGBGradientPalettePtr gGradientPalettes[] = {
 
 
 // Count of how many cpt-city gradients are defined:
-const uint8_t gGradientPaletteCount =
+const uint16_t gGradientPaletteCount =
   sizeof( gGradientPalettes) / sizeof( TProgmemRGBGradientPalettePtr );
