+++ TwinkleFOX.h
@@ -63,12 +63,12 @@
 // Overall twinkle speed.
 // 0 (VERY slow) to 8 (VERY fast).
 // 4, 5, and 6 are recommended, default is 4.
-uint8_t twinkleSpeed = 4;
+uint16_t twinkleSpeed = 4;
 
 // Overall twinkle density.
 // 0 (NONE lit) to 8 (ALL lit at once).
 // Default is 5.
-uint8_t twinkleDensity = 5;
+uint16_t twinkleDensity = 5;
 
 // Background color for 'unlit' pixels
 // Can be set to CRGB::Black if desired.
@@ -99,7 +99,7 @@ CRGBPalette16 twinkleFoxPalette;
 //  /             \
 //
 
-uint8_t attackDecayWave8( uint8_t i)
+uint16_t attackDecayWave8( uint16_t i)
 {
   if( i < 86) {
     return i * 3;
@@ -112,11 +112,11 @@ uint8_t attackDecayWave8( uint8_t i)
 // This function takes a pixel, and if its in the 'fading down'
 // part of the cycle, it adjusts the color a little bit like the
 // way that incandescent bulbs fade toward 'red' as they dim.
-void coolLikeIncandescent( CRGB& c, uint8_t phase)
+void coolLikeIncandescent( CRGB& c, uint16_t phase)
 {
   if( phase < 128) return;
 
-  uint8_t cooling = (phase - 128) >> 4;
+  uint16_t cooling = (phase - 128) >> 4;
   c.g = qsub8( c.g, cooling);
   c.b = qsub8( c.b, cooling * 2);
 }
@@ -130,21 +130,21 @@ void coolLikeIncandescent( CRGB& c, uint8_t phase)
 //  of one cycle of the brightness wave function.
 //  The 'high digits' are also used to determine whether this pixel
 //  should light at all during this cycle, based on the twinkleDensity.
-CRGB computeOneTwinkle( uint32_t ms, uint8_t salt)
+CRGB computeOneTwinkle( uint32_t ms, uint16_t salt)
 {
   uint16_t ticks = ms >> (8-twinkleSpeed);
-  uint8_t fastcycle8 = ticks;
+  uint16_t fastcycle8 = ticks;
   uint16_t slowcycle16 = (ticks >> 8) + salt;
   slowcycle16 += sin8( slowcycle16);
   slowcycle16 =  (slowcycle16 * 2053) + 1384;
-  uint8_t slowcycle8 = (slowcycle16 & 0xFF) + (slowcycle16 >> 8);
+  uint16_t slowcycle8 = (slowcycle16 & 0xFF) + (slowcycle16 >> 8);
 
-  uint8_t bright = 0;
+  uint16_t bright = 0;
   if( ((slowcycle8 & 0x0E)/2) < twinkleDensity) {
     bright = attackDecayWave8( fastcycle8);
   }
 
-  uint8_t hue = slowcycle8 - salt;
+  uint16_t hue = slowcycle8 - salt;
   CRGB c;
   if( bright > 0) {
     c = ColorFromPalette( twinkleFoxPalette, hue, bright, NOBLEND);
@@ -180,7 +180,7 @@ void drawTwinkles()
   if( (AUTO_SELECT_BACKGROUND_COLOR == 1) &&
       (twinkleFoxPalette[0] == twinkleFoxPalette[1] )) {
     bg = twinkleFoxPalette[0];
-    uint8_t bglight = bg.getAverageLight();
+    uint16_t bglight = bg.getAverageLight();
     if( bglight > 64) {
       bg.nscale8_video( 16); // very bright, so scale to 1/16th
     } else if( bglight > 16) {
@@ -192,25 +192,25 @@ void drawTwinkles()
     bg = gBackgroundColor; // just use the explicitly defined background color
   }
 
-  uint8_t backgroundBrightness = bg.getAverageLight();
+  uint16_t backgroundBrightness = bg.getAverageLight();
 
-  for(uint8_t i = 0; i < NUM_LEDS; i++) {
+  for(uint16_t i = 0; i < NUM_LEDS; i++) {
     CRGB& pixel = leds[i];
 
     PRNG16 = (uint16_t)(PRNG16 * 2053) + 1384; // next 'random' number
     uint16_t myclockoffset16= PRNG16; // use that number as clock offset
     PRNG16 = (uint16_t)(PRNG16 * 2053) + 1384; // next 'random' number
     // use that number as clock speed adjustment factor (in 8ths, from 8/8ths to 23/8ths)
-    uint8_t myspeedmultiplierQ5_3 =  ((((PRNG16 & 0xFF)>>4) + (PRNG16 & 0x0F)) & 0x0F) + 0x08;
+    uint16_t myspeedmultiplierQ5_3 =  ((((PRNG16 & 0xFF)>>4) + (PRNG16 & 0x0F)) & 0x0F) + 0x08;
     uint32_t myclock30 = (uint32_t)((clock32 * myspeedmultiplierQ5_3) >> 3) + myclockoffset16;
-    uint8_t  myunique8 = PRNG16 >> 8; // get 'salt' value for this pixel
+    uint16_t  myunique8 = PRNG16 >> 8; // get 'salt' value for this pixel
 
     // We now have the adjusted 'clock' for this pixel, now we call
     // the function that computes what color the pixel should be based
     // on the "brightness = f( time )" idea.
     CRGB c = computeOneTwinkle( myclock30, myunique8);
 
-    uint8_t cbright = c.getAverageLight();
+    uint16_t cbright = c.getAverageLight();
     int16_t deltabright = cbright - backgroundBrightness;
     if( deltabright >= 32 || (!bg)) {
       // If the new pixel is significantly brighter than the background color,
